COBOL Code for use in Projects Honoring Grace Hopper.
In honor of Rear Admiral Grace Hopper, computer pioneer and inventor
of the compiler, I learned enough COBOL over a weekend to write a simple
program that does some rudimentary statistics.

The name of the program, ACTUAL-BUG-INSTANCES, refers to the famous
joke she made at Harvard, when a moth caught in a relay caused a computer
failure and she taped the unfortunate insect into her logbook with the
note, "First actual case of bug being found."

You can see a picture of the log entry (and the moth!) here:
http://www.computerhistory.org/tdih/September/9/

I'm releasing this freely so that anyone who wants to include some COBOL
code in a design honoring Admiral Hopper can draw on it.

Jacob Rose
Brooklyn, New York
September 17, 2016